Moving methods to convert Xamarin.Forms.View to native views to ViewExtensions.cs
