--- Xamarin.Forms.Platform.Android/CarouselView/FormsToNativeDroid.cs
@@ -1,24 +1,0 @@
-﻿
-using Android.Views;
-
-namespace Xamarin.Forms.Platform.Android
-{
-    public static class FormsToNativeDroid
-    {
-        public static ViewGroup ConvertFormsToNative(Xamarin.Forms.View view, Rectangle size)
-        {
-			if (Platform.GetRenderer(view) == null)
-				Platform.SetRenderer(view, Platform.CreateRenderer(view));
-			var vRenderer = Platform.GetRenderer(view);
-            
-            var viewGroup = vRenderer.ViewGroup;
-            vRenderer.Tracker.UpdateLayout ();
-            var layoutParams = new ViewGroup.LayoutParams ((int)size.Width, (int)size.Height);
-            viewGroup.LayoutParameters = layoutParams;
-            view.Layout (size);
-            viewGroup.Layout (0, 0, (int)view.WidthRequest, (int)view.HeightRequest);
-            return viewGroup; 
-        }
-    }
-}
-
