+++ Xamarin.Forms.Platform.Android/Xamarin.Forms.Platform.Android.csproj
@@ -250,7 +250,6 @@
     <Compile Include="CarouselView\CirclePageIndicator.cs" />
     <Compile Include="CarouselView\CustomViewPager.cs" />
     <Compile Include="CarouselView\DefaultPageTransformer.cs" />
-    <Compile Include="CarouselView\FormsToNativeDroid.cs" />
     <Compile Include="CarouselView\PageIndicator.cs" />
     <Compile Include="CarouselView\VerticalViewPager.cs" />
     <Compile Include="AppCompat\Resource.cs" />
