--- Xamarin.Forms.Platform.iOS/CarouselView/FormsViewToNativeiOS.cs
@@ -1,30 +1,0 @@
-﻿
-using UIKit;
-using CoreGraphics;
-
-namespace Xamarin.Forms.Platform.iOS
-{
-	public static class FormsViewToNativeiOS
-	{
-		public static UIView ConvertFormsToNative(Xamarin.Forms.View view, CGRect size)
-		{
-			if (Platform.GetRenderer(view) == null)
-				Platform.SetRenderer(view, Platform.CreateRenderer(view));
-			var vRenderer = Platform.GetRenderer(view);
-
-			vRenderer.NativeView.Frame = size;
-
-			vRenderer.NativeView.AutoresizingMask = UIViewAutoresizing.All;
-			vRenderer.NativeView.ContentMode = UIViewContentMode.ScaleToFill;
-
-			vRenderer.Element.Layout(size.ToRectangle());
-
-			var nativeView = vRenderer.NativeView;
-
-			nativeView.SetNeedsLayout();
-
-			return nativeView;
-		}
-	}
-}
-
