+++ Xamarin.Forms.Platform.iOS/Xamarin.Forms.Platform.iOS.csproj
@@ -153,7 +153,6 @@
     <Compile Include="NativeValueConverterService.cs" />
     <Compile Include="NativeBindingService.cs" />
     <Compile Include="CarouselView\CarouselViewRenderer.cs" />
-    <Compile Include="CarouselView\FormsViewToNativeiOS.cs" />
     <Compile Include="CarouselView\ViewContainer.cs" />
   </ItemGroup>
   <ItemGroup>
